Implement react-graphjs-2
Install react-graphjs-2 and chart.js. Generate labels and data for each
label with two helped methods. Use bitcoin data state property as the
argument for both helpers to generate two different array, one for
grabbing the property key date, and one to grab the property value
prices. My graph labels are the date and the data inside the graph are
prices for each date.